Condense processor shutdown tests
